requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
Django==4.0.1
djangorestframework==3.13.1
google==3.0.0
gunicorn==20.1.0
pytz==2021.3
soupsieve==2.3.1
sqlparse==0.4.2
